// src/app/core/service/model/department.model.ts
export class Department {
    id: number;
    name: string;
    nameDisplay: string;
    code:string;
    parentId: number;
    position: number;
    employeeId:number;
    employeeName: number;
    type: string;
    description: string;
    createdName: string;
    updatedName: string;
    typeImport: number;
    fileUploads: FormData;
    typeSearch: number;
    isLeaf: boolean
}